fix!: rename color tokens
Establish patterns for naming tokens.

Semantic tokens are named by type, category, value, [modifier]
Component tokens are element, type, property, [modifier]

Note that color always refers to the type of token, not the html element property of the same name.

Semantic token example:

`color-background-primary` - color is the token type, background is the token category, and primary is the token value.

Component token example:

`table-color-background` - table is the component, color is the type, and background is the property.

another component example:

`input-color-outline-focus` - input is the component, color is the type, outline is the property, and focus is the modifier.

another component example:

`table-row-color-background-odd` - table-row is the component, color is the type, background is the property, and odd is the modifier.